+++ CHANGELOG.md
@@ -11,6 +11,8 @@
   * ajout `acces-aux-droits-et-citoyennete--aide-aux-victimes`
   * ajout des thématiques liées à l'illettrisme (`illettrisme--*`)
 * ajouts des types de zone de diffusion `pays` et `epci`
+* profils:
+  * ajout `beneficiaire-rsa`
 
 ## Version 0.8.0 - 2023-03-08